services/api/requirements.txt
duckdb>=1.1.0
fastapi>=0.115.0
openai>=1.75.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
pyyaml>=6.0.2
sqlglot>=25.20.0
uvicorn[standard]>=0.32.0
